Move the logic of Corfu table creation in a utility file.
